Refactor agent chat message trigger logic into helper functions
